feat: add auth service
