+++ src/middlewares/auth.ts
@@ -1,9 +1,9 @@
 /* eslint-disable import/no-unresolved */
 import { Request, Response, NextFunction } from 'express';
-import * as admin from 'firebase-admin';
 import { UserRecord } from 'firebase-admin/auth';
 
 import config from '../config';
+import auth from '../services/auth';
 
 export interface AuthenticatedRequest extends Request {
   user?: UserRecord;
@@ -13,22 +13,6 @@ const devUser = {
   uid: 'PEU8bBVrBxa2RzjklhARkCW6ulY2',
   email: 'dev@fiuber.uba.ar',
 };
-
-const firebaseConfig = {
-  projectId: config.firebase.projectId,
-  privateKey: config.firebase.privateKey,
-  clientEmail: config.firebase.clientEmail,
-};
-
-if (!admin.apps.length) {
-  /** See https://stackoverflow.com/a/57764002/2516673 */
-  admin.initializeApp({
-    projectId: firebaseConfig.projectId,
-    credential: admin.credential.cert(firebaseConfig),
-  });
-}
-
-const auth = admin.auth();
 
 const firebaseAuth = async (
   req: AuthenticatedRequest,
